+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 ### v3.10.2 / 2024-04
 
+* [改进 bilibili 视频网站解析](https://github.com/Vanessa219/vditor/issues/1581) `改进功能`
 * [改进松散模式下的列表展现](https://github.com/Vanessa219/vditor/issues/1575) `改进功能`
 * [mermaid 图表渲染自适应](https://github.com/Vanessa219/vditor/issues/1574) `改进功能`